+++ packages/atlas/lib/atlas.dart
@@ -16,4 +16,7 @@ export 'src/polygon.dart';
 export 'src/polyline.dart';
 export 'src/screen_coordinate.dart';
 export 'src/poi.dart';
+export 'src/bounding_box_data.dart';
+export 'src/rectangle_2d.dart';
+export 'src/move_camera_animation.dart';
 export 'src/annotation_type.dart';