+++ packages/atlas/lib/src/atlas_controller.dart
@@ -4,20 +4,17 @@ import 'package:atlas/atlas.dart';
 
 /// `AtlasController` allows developers to manipulate `Atlas` after the map has been initialized.
 abstract class AtlasController {
-  /// Moves the camera to the specified `CameraPosition`
-  Future<void> moveCamera(CameraPosition cameraPosition);
+  /// Moves the camera to the specified `CameraPosition` with an optional animation `​​​​​MoveCameraAnimation`.
+  Future<void> moveCamera(
+    CameraPosition cameraPosition, {
+    MoveCameraAnimation animation,
+  });
 
   /// Get the current camera position
   Future<CameraPosition> getCameraPosition();
 
-  /// Updates the camera bounds based on the provided `LatLngBounds`
-  /// and adds padding around the bounds based on the provided `padding`.
-  Future<void> updateBounds(LatLngBounds bounds, double padding);
-
-  /// Updates the camera bounds based on the provided `LatLngBounds`
-  /// and adds padding around the bounds based on the provided `north, east, south, west` padding.
-  Future<void> updateBoundsWithPaddingToAllSides(LatLngBounds bounds,
-      double north, double east, double south, double west);
+  /// Updates the camera bounds based on the provided `BoundingBoxData`
+  Future<void> updateBounds(BoundingBoxData boundingBoxData);
 
   /// Return a [ScreenCoordinate] corresponding to the [LatLng] in the map view.
   Future<LatLng> getLatLng(ScreenCoordinates screenCoordinates);